Task 2 nearly done

+++ src/main/java/org/launchcode/techjobs/persistent/models/data/EmployerRepository.java
@@ -1,9 +1,10 @@
 package org.launchcode.techjobs.persistent.models.data;
 
+import org.springframework.data.repository.CrudRepository;
 import org.launchcode.techjobs.persistent.models.Employer;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployerRepository extends CrudRepository<Employer, Long> {
+public interface EmployerRepository extends CrudRepository<Employer, Integer> {
 
 }
